// frontend/src/types.ts
type ColorType = RGB | RGBA | HEX;

interface payload {
  paths: string[];
  position: { x: number; y: number };
  id: number;
}

type Task = { text: string; check: boolean; nestLevel: number };

enum NodeType {
  Note = "Note",
  Comment = "Comment",
  Todo = "Todo",
  Table = "Table",
  Url = "Url",
  Activity = "Activity",

  Arrow = "Arrow",
  Board = "Board",
  Column = "Column",

  Code = "Code",
  Document = "Document",
  Upload = "Upload",

  Drawing = "Drawing",
  Sketch = "Sketch",
  Color = "Color",
  Image = "Image",
}

interface BaseNode {
  id: string;
  type: NodeType;
}

interface Node extends BaseNode {
  x: number;
  y: number;
  //? can be decided by the user
  width?: number;
  //? height depends on content shouldnt be user controlled
  //height?: number;
  color?: ColorType;
  textColor?: ColorType;
  top_strip_color?: ColorType;
  zIndex?: number;
}

interface ChildNode extends BaseNode {
  id: string;
  type: NodeType;
  index: number;
  color?: ColorType;
}

interface Note extends Node, ChildNode {
  text: string; // rich text maybe json
}

interface Comment extends Node, ChildNode {
  comment: string;
}

interface Url extends Node, ChildNode {
  url: string;
}
interface Todo extends Node, ChildNode {
  title?: string;
  // [ ] text
  // [x] text
  tasks: Task[];
}

////////////////////////////////////////////////

type Badge = {
  id: string;
  label: string;
  color: ColorType;
};

type BadgeRegistry = Record<string, Badge[]>;

enum ColumnType {
  String = "String",
  Number = "Number",
  Boolean = "Boolean",
  Date = "Date",
  Badge = "Badge",
}
// can add more in the future

interface ColumnSchema {
  key: string;
  title: string;
  typeDef: ColumnType;
}
interface Table extends Node, ChildNode {
  columns: ColumnSchema[];
  rows: Record<string, string | number | Badge>[];
  description?: string;
  title?: string;
}
////////////////////////////////////////////////

interface Column extends Node, ChildNode {
  title: string;
}

interface Color extends Node, ChildNode {
  colorValue: ColorType;
  text: string;
}

interface Board extends Node, ChildNode {
  name: string;
  icon_path: string;
}
interface Image extends Node, ChildNode {
  path: string;
  title: string;
  description: string;
}
interface Activity extends Node, ChildNode {
  //TODO: figure me out
  title: string;
  desciption: string;
  //? key string is for the date yyyy-mm-dd
  // ! pain in the ass to implment it to handle other date formats
  progress: Record<string, number>;
}

interface Arrow extends Node, ChildNode {}
interface Code extends Node, ChildNode {}
interface Upload extends Node, ChildNode {}
interface Drawing extends Node, ChildNode {}
interface Sketch extends Node, ChildNode {}
interface Document extends Node, ChildNode {
  path: string;
  // widget = like board small round
  // widget = you see file first page, name, download, open, size
  // reader = custom make it a full readable widget
  docType: "widget" | "card" | "reader";
  // appears only in card docType
  description: string;
}

////////////////////////////////////////////////

enum EdgeType {
  Bezier = "Bezier",
  Straight = "Straight",
  Step = "Step",
}

interface BaseEdge {
  srcNodeId: string;
  distNodeId: string;
  color?: string; // default foreground
  stroke?: string; // default 2
  label?: string; // default none
  style?: string; // solid, dashed ...., default solid
  // type?: string; // straight, bezier, step curved >>>> default bezier
  srcArrowHead?: string; // default none
  DistArrowHead?: string; // default normal arrow head
}
interface BezierEdge extends BaseEdge {
  type: EdgeType.Bezier;
  curvature?: number;
}
interface StraightEdge extends BaseEdge {
  type: EdgeType.Straight;
  sharpness?: number;
}
interface StepEdge extends BaseEdge {
  type: EdgeType.Step;
  stepDirection?: "horizontal" | "vertical";
}
type Edge = BezierEdge | StraightEdge | StepEdge;

////////////////////////////////////////////////

type NodeUnion =
  | Note
  | Comment
  | Todo
  | Table
  | Url
  | Activity
  | Arrow
  | Board
  | Column
  | Code
  | Document
  | Upload
  | Drawing
  | Sketch
  | Color
  | Image;